Move TracklistItem component into organisms
