+++ src/pages/Tracklists/Index.tsx
@@ -8,7 +8,7 @@ import API, { Tracklist } from "services/memoir";
 import Footer from "components/Footer";
 import Loading from "components/Loading";
 import Pagination from "components/Pagination";
-import TracklistItem from "components/TracklistItem";
+import TracklistItem from "components/organisms/TracklistItem";
 
 const className = css({
   minHeight: "33.625rem",
@@ -54,7 +54,16 @@ export default ({ path, page, api }: Props) => {
     <div class={className}>
       {loading && <Loading />}
       {tracklists &&
-        tracklists.map((tracklist) => <TracklistItem tracklist={tracklist} />)}
+        tracklists.map(({ id, name, date, artwork, trackCount }) => (
+          <TracklistItem
+            key={id}
+            id={id}
+            name={name}
+            date={date}
+            artwork={artwork}
+            trackCount={trackCount}
+          />
+        ))}
       <Pagination path={path!} page={parseInt(page!, 10)} hasMore={hasMore} />
       <Footer />
     </div>
